recipe_crawler_simple/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
filelock==3.18.0
frozenlist==1.5.0
google-auth==2.38.0
google-genai==1.7.0
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.1
idna==3.10
lxml==5.3.1
multidict==6.2.0
propcache==0.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
tenacity==9.0.0
tldextract==5.1.3
typing_extensions==4.12.2
urllib3==2.3.0
websockets==15.0.1
yarl==1.18.3
